rabbit_prelaunch_errors: Handle exception stacktraces with args list
... instead of function arity.

I don't know when this was introduced, perhaps Erlang 23. Anyway, each
stacktrace entry has now the form:

    {Mod, Fun, ArgListOrArity, Props}

where `ArgListOrArity` is either the function arity (an integer) or the
actual list of arguments passed to the function.

If we get the latter, we format the stacktrace line as usual and we add
the list of arguments to the line below it.